src/test/java/br/com/fiap/lanchonete/pedidoservicefase4/infra/db/postgresql/PedidoPostgresqlRepositoryTest.java
package br.com.fiap.lanchonete.pedidoservicefase4.infra.db.postgresql;

import br.com.fiap.lanchonete.pedidoservicefase4.domain.entities.OrdemPedido;
import br.com.fiap.lanchonete.pedidoservicefase4.domain.entities.Pedido;
import br.com.fiap.lanchonete.pedidoservicefase4.domain.enums.StatusEnum;
import br.com.fiap.lanchonete.pedidoservicefase4.infra.db.entities.PedidoEntity;
import br.com.fiap.lanchonete.pedidoservicefase4.infra.db.mappers.PedidoDataMapper;
import br.com.fiap.lanchonete.pedidoservicefase4.infra.db.postgresql.PedidoPostgresqlRepository;
import br.com.fiap.lanchonete.pedidoservicefase4.infra.db.repository.PedidoRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.modelmapper.ModelMapper;

import java.util.Arrays;
import java.util.List;
import java.util.Objects;
import java.util.Optional;
import java.util.stream.Collectors;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.Mockito.*;

public class PedidoPostgresqlRepositoryTest {

    @InjectMocks
    private PedidoPostgresqlRepository pedidoPostgresqlRepository;

    @Mock
    private PedidoRepository pedidoRepository;

    @Mock
    private ModelMapper modelMapper;

    @Mock
    private PedidoDataMapper pedidoDataMapper;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    public void getPedidoById() {
        PedidoEntity pedidoEntity = new PedidoEntity();
        Pedido pedido = new Pedido();
        when(pedidoRepository.findById(anyLong())).thenReturn(Optional.of(pedidoEntity));
        when(modelMapper.map(pedidoEntity, Pedido.class)).thenReturn(pedido);

        Pedido result = pedidoPostgresqlRepository.get(1L);

        assertNotNull(result);
        assertEquals(pedido, result);
    }

    @Test
    public void findAllPedidos() {
        List<PedidoEntity> pedidoEntities = Arrays.asList(new PedidoEntity(), new PedidoEntity());
        List<Pedido> pedidos = Arrays.asList(new Pedido(), new Pedido());
        when(pedidoRepository.findAll()).thenReturn(pedidoEntities);
        when(modelMapper.map(pedidoEntities.get(0), Pedido.class)).thenReturn(pedidos.get(0));
        when(modelMapper.map(pedidoEntities.get(1), Pedido.class)).thenReturn(pedidos.get(1));

        List<Pedido> result = pedidoPostgresqlRepository.findAll();

        assertNotNull(result);
        assertEquals(2, result.size());
    }



    @Test
    public void savePedido() {
        Pedido pedido = new Pedido();
        PedidoEntity pedidoEntity = new PedidoEntity();
        when(pedidoDataMapper.toData(any(Pedido.class))).thenReturn(pedidoEntity);
        when(pedidoRepository.save(any(PedidoEntity.class))).thenReturn(pedidoEntity);
        when(modelMapper.map(pedidoEntity, Pedido.class)).thenReturn(pedido);

        Pedido result = pedidoPostgresqlRepository.save(pedido);

        assertNotNull(result);
        assertEquals(pedido, result);
    }

    @Test
    public void checkoutPedido() {
        Pedido pedido = new Pedido();
        PedidoEntity pedidoEntity = new PedidoEntity();
        when(modelMapper.map(any(Pedido.class), eq(PedidoEntity.class))).thenReturn(pedidoEntity);
        when(pedidoRepository.save(any(PedidoEntity.class))).thenReturn(pedidoEntity);

        Pedido result = pedidoPostgresqlRepository.checkout(pedido);

        assertNotNull(result);
        assertEquals(pedido, result);
    }






}
